+++ app/layout.tsx
@@ -1,12 +1,15 @@
-import '../styles/globals.css';
-import ReactQueryWrapper from './ReactQueryWrapper';
+import "@/styles/globals.css";
+import ReactQueryWrapper from "./ReactQueryWrapper";
+import { Raleway } from "next/font/google";
+
+const raleway = Raleway({
+  subsets: ["latin"],
+  variable: "--font-raleway",
+});
 
 const RootLayout = ({ children }: { children: React.ReactNode }) => (
-  <html lang='en'>
-    <head>
-      <link href='https://fonts.googleapis.com/css2?family=Raleway:wght@400;700&display=swap' rel='stylesheet' />
-    </head>
-    <body className='font-raleway'>
+  <html lang="en" className={raleway.variable}>
+    <body>
       <ReactQueryWrapper>{children}</ReactQueryWrapper>
     </body>
   </html>
