+++ components/Grid/Grid.tsx
@@ -6,8 +6,8 @@ type Props = {
 
 const Grid = ({ title, children, className }: Props) => (
   <div className={className}>
-    <h2 className='text-xl font-bold pb-4'>{title}</h2>
-    <div className='grid grid-cols-auto-fill gap-8'>{children}</div>
+    <h2 className="text-xl font-bold pb-4">{title}</h2>
+    <div className="grid grid-cols-auto-fill gap-8">{children}</div>
   </div>
 );
