+++ next.config.js
@@ -1,13 +1,13 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
   images: {
-    domains: ['image.tmdb.org']
-  },
-  experimental: {
-    appDir: true,
-    // enableUndici: true   use this if Node < v18
+    remotePatterns: [
+      {
+        protocol: "http",
+        hostname: "image.tmdb.org",
+        pathname: "/t/p/**",
+      },
+    ],
   },
 };
