+++ tailwind.config.js
@@ -11,7 +11,7 @@ module.exports = {
       }
     },
     fontFamily: {
-      raleway: ['Raleway', 'sans-serif']
+      sans: ['var(--font-raleway)', 'sans-serif']
     },
     minHeight: {
       128: '40rem'
